chore: deploy staking bank static on prod

+++ deploy/evm/UmbrellaFeeds.deploy.ts
@@ -8,6 +8,8 @@ import { umbrellaFeedsDeploymentData } from '../deploymentsData/umbrellaFeeds';
 import { checkStakingBankStaticUpdated } from '../_helpers/checkStakingBankStaticUpdated';
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
+  return;
+
   const { deployments, getNamedAccounts } = hre;
   const { deploy, read } = deployments;
   const { deployer } = await getNamedAccounts();
@@ -19,18 +21,22 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
     return;
   }
 
-  const [bankInFees, recentBank] = await Promise.all([
-    read(UMBRELLA_FEEDS, 'STAKING_BANK'),
-    deployments.get(STAKING_BANK_STATIC),
-  ]);
+  const f = `${__dirname}/../deployments/${hre.network.name}/${UMBRELLA_FEEDS}.json`;
 
-  if (bankInFees.toLowerCase() == recentBank.address.toLowerCase()) {
-    console.log(`${UMBRELLA_FEEDS} has current bank ${recentBank.address} - OK`);
-  } else {
-    console.log('new bank detected - deploying...');
-    const f = `${__dirname}/../deployments/${hre.network.name}/${UMBRELLA_FEEDS}.json`;
-    console.log('new bank detected - deploying...', f);
-    fs.unlinkSync(f);
+  if (fs.existsSync(f)) {
+    const [bankInFees, recentBank] = await Promise.all([
+      read(UMBRELLA_FEEDS, 'STAKING_BANK'),
+      deployments.get(STAKING_BANK_STATIC),
+    ]);
+
+    if (bankInFees.toLowerCase() == recentBank.address.toLowerCase()) {
+      console.log(`${UMBRELLA_FEEDS} has current bank ${recentBank.address} - OK`);
+    } else {
+      console.log('new bank detected - deploying...');
+      const f = `${__dirname}/../deployments/${hre.network.name}/${UMBRELLA_FEEDS}.json`;
+      console.log('new bank detected - deploying...', f);
+      fs.unlinkSync(f);
+    }
   }
 
   const feeds = await deploy(UMBRELLA_FEEDS, {
